+++ src/main/java/com/everyonewaiter/global/support/PagingRequest.java
@@ -1,22 +1,22 @@
 package com.everyonewaiter.global.support;
 
-public record PagingRequest(int page, int size, int pageSkipSize) {
+public record PagingRequest(long page, long size, int pageSkipSize) {
 
   private static final int DEFAULT_PAGE_SKIP_SIZE = 5;
 
-  public PagingRequest(int page, int size) {
+  public PagingRequest(long page, long size) {
     this(page, size, DEFAULT_PAGE_SKIP_SIZE);
   }
 
-  public int countLimit() {
+  public long countLimit() {
     return ((page + pageSkipSize - 1) * size) + 1;
   }
 
-  public int limit() {
+  public long limit() {
     return size;
   }
 
-  public int offset() {
+  public long offset() {
     return (page - 1) * size;
   }
